requirements.txt
astronomer-cosmos[dbt-bigquery]==1.0.3
soda-core-bigquery==3.0.45
protobuf